pyproject.toml
[project]
name = "pawpaw"
version = "0.1.0"
description = "MLOps project for pet pawpularity prediction"
requires-python = "==3.12.*"
dependencies = [
    "boto3==1.39.11",
    "boto3-stubs[essential]==1.39.11",
    "fastapi[standard]==0.116.1",
    "ipykernel>=6.30.1",
    "kaleido==0.2.1",
    "lakefs==0.12.0",
    "lightning==2.5.2",
    "mlflow==3.2.0",
    "nannyml==0.13.1",
    "nbformat==5.10.4",
    "pillow==11.3.0",
    "plotly==5.24.1",
    "polars==1.32.2",
    "prefect>=3.4.12",
    "psutil==7.0.0",
    "psycopg2==2.9.10",
    "pydantic==2.11.7",
    "pynvml==12.0.0",
    "python-multipart==0.0.20",
    "sqlalchemy==2.0.43",
    "streamlit==1.48.0",
    "torch==2.8.0",
    "torchinfo==1.8.0",
    "torchvision==0.23.0",
]

[tool.uv.sources]
torch = [ { index = "pytorch-cu128", marker = "sys_platform != 'darwin'" } ]
torchvision = [ { index = "pytorch-cu128", marker = "sys_platform != 'darwin'" } ]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[build-system]
requires = ["uv_build==0.8.9"]
build-backend = "uv_build"
